requirements.txt
python-dotenv==1.0.1
streamlit==1.36.0
openai==1.43.0
pymupdf==1.24.9
python-docx==1.1.2
tiktoken==0.7.0